% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cope.R
\name{cope}
\alias{cope}
\title{covariance parameter estimates with fast methods for big data with SSN}
\usage{
cope(formula, ssn.object, CorModels = c("Exponential.tailup",
  "Exponential.taildown", "Exponential.Euclid"), use.nugget = TRUE,
  addfunccol = NULL, EstMeth = "REML", partIndxCol, parallel = FALSE)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be 
coerced to that class): a symbolic description of the model to be}

\item{ssn.object}{an object of class
\code{SpatialStreamNetwork}, 
representing a spatial stream network. This contains the variables 
used in the model.}

\item{CorModels}{a vector of spatial autocorrelation models for 
stream networks.}

\item{use.nugget}{add a nugget effect, default is TRUE. This can be 
thought of as a variance component for independent errors, adding 
a variance component only along the diagonal of the covariance matrix.}

\item{addfunccol}{the name of the variable in the 
SpatialStreamNetwork object that is used to define spatial 
weights. For the tailup models, weights need to be used for 
branching. This is an additive function and is described in 
Ver Hoef and Peterson (2010). See example below.}

\item{EstMeth}{Estimation method; either "ML" for maximum 
likelihood, or "REML" for restricted maximum likelihood (default).}

\item{subSampIndxCol}{the column in the \code{points data.frame}
within the \code{SpatialStreamNetwork} object that indexes the
grouping to be used when subsampling.}
}
\value{
an object of class "estCovParSSNbd", which is a list, where 
  \code{estCovPar} is a vector of estimated covariance parameters, 
  \code{optimOut} is the output from \code{optim} used to estimate
  the parameters.
}
\description{
covariance parameter estimates with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
