% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFixEffSSNbd.R
\name{estFixEffSSNbd}
\alias{estFixEffSSNbd}
\title{estimates fixed effects with fast methods for big data with SSN}
\usage{
estFixEffSSNbd(estCovParSSN)
}
\arguments{
\item{estCovParSSN}{an object of class "estCovParSSNbd" created
from estimating the covariance parameters.}
}
\value{
a list
}
\description{
estimates fixed effects with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
