% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFixEffSSNbd_orig.R
\name{estFixEffSSNbd_orig}
\alias{estFixEffSSNbd_orig}
\title{estimates fixed effects with fast methods for big data with SSN}
\usage{
estFixEffSSNbd_orig(estCovParSSN)
}
\arguments{
\item{estCovParSSN}{an object of class "estCovParSSNbd" created
from estimating the covariance parameters.}
}
\value{
a list
}
\description{
estimates fixed effects with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
