% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fefe.R
\name{fefe}
\alias{fefe}
\title{fixed effects estimation with fast methods for big data with SSN}
\usage{
fefe(estCovParSSN)
}
\arguments{
\item{estCovParSSN}{an object of class "estCovParSSNbd" created
from estimating the covariance parameters.}
}
\value{
a list
}
\description{
fixed effects estimates with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
