% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreamDistMatInt.R
\name{getStreamDistMatInt}
\alias{getStreamDistMatInt}
\title{Extracts downstream-only hydrologic distances for two vectors of pids}
\usage{
getStreamDistMatInt(ssn, pidset1, dataID1, pidset2 = NULL, dataID2 = NULL)
}
\arguments{
\item{ssn}{a SpatialStreamNetwork object created using the SSN package}

\item{pidset1}{a vector of pids found within a SpatialStreamNetwork object}

\item{dataID1}{a character string representing the name of the dataset found
in the SpatialStreamNetwork object slot ID}

\item{pidset2}{a vector of pids found within a SpatialStreamNetwork object.
Default = NULL.}

\item{dataID2}{a character string representing the name of the dataset found
in the SpatialStreamNetwork object slot ID. Default = NULL.}
}
\value{
a list matrices containing the of downstream-only hydrologic distance for
  pidset1 and pidset2. If pidset2 is omitted, the function returns the downstream-only
  hydrologic distances between pidset1 and pidset1. See documentation for
  createBigDistMat for more information about the format of the distance matrices.
}
\description{
Extracts downstream-only hydrologic distances for two vectors of pids and
returns a list of distance matrices.
}
\examples{
## ssn<- importSSN("raw5.ssn", predpts = "preds")
## createBigDistMat(ssn, predpts = "preds", o.write = TRUE, amongpreds = TRUE,
##                  no.cores = 2)
## pidset1 <- c(1, 3, 9, 28, 34, 51)
## pidset2 <- c(4, 7, 30, 35)
## distmats <- getStreamDistMatInt(ssn, pidset1, dataID1 = "preds",
##                                  pidset2 = pidset2, dataID2 = "preds")
}
\author{
Erin Peterson
}
