% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapi.R
\name{mapi}
\alias{mapi}
\title{make a partition index for a SpatialStreamNetwork object}
\usage{
mapi(ssn.object, nIndx, partMeth = "comp", set.seed.no = 1001)
}
\arguments{
\item{ssn.object}{a SpatialStreamNetwork object, usually imported 
with SSN package}

\item{nIndx}{the number of indexes (number of groups for partitioning)}

\item{partMeth}{method for subsampling.  Options are 'rand' for 
random, 'comp' for compact blocks using kmeans on coordinates,
and 'zimm' for compact blocks with 10% re-assigned at random.
Default is 'zimm'.}

\item{set.seed.no}{set the random number seed so results are repeatable.}
}
\value{
an object of class "SpatialStreamNetwork", which is the same 
  as the input ssn.object, except the points data.frame within now
  has a new field named partIndx which indexes the groups for 
  partitioning.
}
\description{
make a partition index for a SpatialStreamNetwork object so that 
  it can be used with fast methods for big data
}
\author{
Jay Ver Hoef
}
