% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masi.R
\name{masi}
\alias{masi}
\title{make a subsample index for a SpatialStreamNetwork object}
\usage{
masi(ssn.object, nIndx)
}
\arguments{
\item{ssn.object}{a SpatialStreamNetwork object, usually imported 
with SSN package}

\item{nIndx}{the number of indexes (number of groups for subsampling)}
}
\value{
an object of class "SpatialStreamNetwork", which is the same 
  as the input ssn.object, except the points data.frame within now
  has a new field named subSampIndx which indexes the groups for 
  subsampling.
}
\description{
make a subsample index for a SpatialStreamNetwork object so that 
  it can be used with fast methods for big data
}
\author{
Jay Ver Hoef
}
