% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predPtsSSNbd.R
\name{predSSNbdNN}
\alias{predSSNbdNN}
\title{predictions with fast methods for big data with SSN}
\usage{
predSSNbdNN(ecp, efe, predsID, nNN)
}
\arguments{
\item{ecp}{object from estCovParSSNbd function.}

\item{efe}{object from estFixEffSSNbd function}

\item{predsID}{name of prediction data set in ssn object (passed with ecp)}

\item{nNN}{number of nearest neighbors for predictions}
}
\value{
a data.frame with predictions in first column, and prediction standard errors in the second column.
}
\description{
Predictions fixed effects with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
