% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estCovParSSNbd.R
\name{print.estCovParSSNbd}
\alias{print.estCovParSSNbd}
\title{Print method for estCovParSSNbd}
\usage{
\method{print}{estCovParSSNbd}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Print method for estCovParSSNbd
}
