% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estFixEffSSNbd.R
\name{print.estFixEffSSNbd}
\alias{print.estFixEffSSNbd}
\title{Print method for estFixEffSSNbd}
\usage{
\method{print}{estFixEffSSNbd}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Print method for estFixEffSSNbd
}
