% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estCovParSSNbd.R
\name{print.summary.estCovParSSNbd}
\alias{print.summary.estCovParSSNbd}
\title{Print method for estCovParSSNbd}
\usage{
\method{print}{summary.estCovParSSNbd}(x, digits = max(3L, getOption("digits")
  - 3L))
}
\arguments{
\item{x}{}
}
\description{
Print method for estCovParSSNbd
}
