% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estFixEffSSNbd.R
\name{print.summary.estFixEffSSNbd}
\alias{print.summary.estFixEffSSNbd}
\title{Print method for estFixEffSSNbd}
\usage{
\method{print}{summary.estFixEffSSNbd}(x,
  signif.stars = getOption("show.signif.stars"), ...)
}
\description{
Print method for estFixEffSSNbd
}
