% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poba1.R
\name{pure}
\alias{pure}
\title{prediction of unobserved responses with fast methods for big data with SSN}
\usage{
pure(ecp, efe, predsID, nNN, poba_prep = FALSE)
}
\arguments{
\item{ecp}{object from cope function.}

\item{efe}{object from fefe function}

\item{predsID}{name of prediction data set in ssn object (passed with ecp)}

\item{nNN}{number of nearest neighbors for predictions}
}
\value{
a data.frame with predictions in first column, and prediction standard errors in the second column.
}
\description{
Prediction of unobserved responses with fast methods for big data with SSN
}
\author{
Jay Ver Hoef
}
