% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estCovParSSNbd.R
\name{summary.estCovParSSNbd}
\alias{summary.estCovParSSNbd}
\title{Summary of S3 Method for Class "estCovParSSNbd"}
\usage{
\method{summary}{estCovParSSNbd}(object, ...)
}
\arguments{
\item{object}{of class \code{estCovParSSNbd}}
}
\value{
covariance parameter estimates for each covariance model.
}
\description{
Summary of S3 Method for Class "estCovParSSNbd"
}
\seealso{
\code{\link{estCovParSSNbd}}
}
\author{
Erin Peterson
}
