% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.estFixEffSSNbd.R
\name{summary.estFixEffSSNbd}
\alias{summary.estFixEffSSNbd}
\title{Summary of S3 Method for Class "estFixEffSSNbd"}
\usage{
\method{summary}{estFixEffSSNbd}(object, ...)
}
\arguments{
\item{object}{of class \code{estFixEffSSNbd}}
}
\value{
Table of estimates for fixed effects.
}
\description{
Summary of S3 Method for Class "estFixEffSSNbd"
}
\seealso{
\code{\link{estFixEffSSNbd}}
}
\author{
Erin Peterson
}
